Clean up Visual C++ lint: 	* client.c (read_line): Change input_index and result_size to size_t. 	(update_entries): Remove unused variables buf2, size_left, size_read. 	(handle_mode): Prototype. 	* client.c, client.h (send_to_server, read_from_server): Change len to size_t. 	* client.c (send_to_server): Change wrtn to size_t. 	(read_from_server): Change red to size_t. 	* client.c, myndbm.c, edit.c, fileattr.c: Include getline.h. 	* checkin.c, commit.c, update.c: Include fileattr.h. 	* commit.c, update.c: Include edit.h. 	* edit.c (onoff_filesdoneproc): Prototype. 	(ncheck_fileproc,edit_fileproc): Change "return" to "return 0". 	(notify_do): Cast a signed value to unsigned before comparing with unsigned.
CVS-ID: src/ChangeLog 1.342
CVS-ID: src/checkin.c 1.14
CVS-ID: src/client.c 1.70
CVS-ID: src/client.h 1.12
CVS-ID: src/commit.c 1.35
CVS-ID: src/edit.c 1.4
CVS-ID: src/fileattr.c 1.2
CVS-ID: src/myndbm.c 1.7
CVS-ID: src/update.c 1.38

+++ src/client.h
@@ -60,8 +60,8 @@ int connect_to_pserver PROTO((int *tofdp, int* fromfdp, int verify_only));
 #endif /* AUTH_CLIENT_SUPPORT */
 
 /* Talking to the server. */
-void send_to_server PROTO((char *str, int len));
-void read_from_server PROTO((char *buf, int len));
+void send_to_server PROTO((char *str, size_t len));
+void read_from_server PROTO((char *buf, size_t len));
 
 /* Internal functions that handle client communication to server, etc.  */
 int supported_request PROTO ((char *));
